requirements.txt
# /requirements.txt
flask
requests
beautifulsoup4
google-generativeai
markdown
google-api-python-client
thefuzz
python-dotenv